requirements.txt
torch
pillow
requests
tqdm
safetensors
git+https://github.com/modelscope/DiffSynth-Studio.git